feat: 대출 상세 조회 시 상환 진척률도 같이 응답되도록 수정

+++ loan-mate/src/main/java/com/fisa/bank/loan/application/dto/response/LoanProgressResponse.java
@@ -5,17 +5,15 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigDecimal;
-
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @Getter
 public class LoanProgressResponse {
   private final Long loanId;
   private final String name;
-  private BigDecimal progress;
+  private Integer progress;
 
-  public static LoanProgressResponse from(Long loanId, String name, BigDecimal progress) {
+  public static LoanProgressResponse from(Long loanId, String name, Integer progress) {
 
     return new LoanProgressResponse(loanId, name, progress);
   }
